Album-Generic-Page
This is a first pass at creating the album page outlined in JoAnna's workup of the wireframe. The lorem ipsum passage is supposed to be an about section, but has been temporarily held since we're just working on the HTML structure. If we're extending the toolbar or if we need to replace the one currently on this page, we can replace/remove the code from lines 68 to 101.