Create drop_down_lyrics_button.html
This is the logic that goes for the lyrics button on the music pages; works fine on desktop but need to make sure it's working on mobile. So far, we have over 30 songs (from Backward Books, Backward Books 2, Xenophile, and I'm Good on People EP); if we need more for the end result, let me know!